refactor(navigation): extract waitForNavigationReady helper from deepLinking saga
