+++ app/lib/navigation/appNavigation.ts
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import { CommonActions, type NavigationContainerRef, StackActions } from '@react-navigation/native';
 
+import { emitter } from '../../methods/helpers';
+
 // TODO: we need change this any to the correctly types from our stacks
 const navigationRef = React.createRef<NavigationContainerRef<any>>();
 const routeNameRef: React.MutableRefObject<NavigationContainerRef<any> | null> = React.createRef();
@@ -72,6 +74,19 @@ function getCurrentRoute() {
 
 function setParams(params: any) {
 	navigationRef.current?.setParams(params);
+}
+
+export function waitForNavigationReady(): Promise<void> {
+	if (navigationRef.current) {
+		return Promise.resolve();
+	}
+	return new Promise(resolve => {
+		const listener = () => {
+			emitter.off('navigationReady', listener);
+			resolve();
+		};
+		emitter.on('navigationReady', listener);
+	});
 }
 
 export default {
